Add distutils

+++ _pymovex.c
@@ -23,15 +23,15 @@
 
 
 static SERVER_ID comStruct;
-static int result;
+static unsigned long result;
 static unsigned long len;
 static PyObject *PyMovexError;
 static char errstr[1024];
 static PyObject*OrderedDict;
 static int DEBUG = 0;
 
-PyObject*reportError(char*method, int result) {
-    snprintf(errstr, sizeof(errstr), "%s returned error %d: %s", method, result, comStruct.Buff);
+PyObject*reportError(char*method, unsigned long result) {
+    snprintf(errstr, sizeof(errstr), "%s returned error %lu: %s", method, result, comStruct.Buff);
     PyErr_SetString(PyMovexError, errstr);
     return NULL;
 }
